+++ services/importer/lib/importer/source_file.rb
@@ -1,6 +1,6 @@
 # encoding: utf-8
 module CartoDB
-  module Importer
+  module Importer2
     class SourceFile
       def initialize(filepath, filename=nil)
         self.filepath = filepath
@@ -26,6 +26,11 @@ module CartoDB
         File.basename(fullpath)
       end #path
 
+      def target_schema
+        return 'public' if extension == '.osm'
+        return 'cdb_importer'
+      end #target_schema
+
       attr_reader :filename
       
       private
@@ -33,6 +38,6 @@ module CartoDB
       attr_writer   :filename
       attr_accessor :filepath
     end # SourceFile
-  end # Importer
+  end # Importer2
 end # CartoDB