+++ services/importer/spec/acceptance/kml_spec.rb
@@ -1,0 +1,53 @@
+# encoding: utf-8
+gem 'minitest'
+require 'minitest/autorun'
+require_relative '../../lib/importer/runner'
+require_relative '../../lib/importer/job'
+require_relative '../../lib/importer/downloader'
+require_relative '../factories/pg_connection'
+
+include CartoDB::Importer2
+
+describe 'KML regression tests' do
+  before do
+    @pg_options  = Factories::PGConnection.new.pg_options
+  end
+
+  it 'imports KML files' do
+    filepath    = path_to('counties_ny_export.kml')
+    downloader  = Downloader.new(filepath)
+    runner      = Runner.new(@pg_options, downloader)
+    runner.run
+
+    geometry_type_for(runner).wont_be_nil
+  end
+
+  it 'imports KML files from url' do
+    filepath    = "https://maps.google.com/maps/ms?hl=en&ie=UTF8&t=m" +
+                  "&source=embed&authuser=0&msa=0&output=kml"         + 
+                  "&msid=214357343079009794152.0004d0322cc2768ca065e"
+    downloader  = Downloader.new(filepath)
+    runner      = Runner.new(@pg_options, downloader)
+    runner.run
+
+    geometry_type_for(runner).wont_be_nil
+  end
+
+  def path_to(filepath)
+    File.expand_path(
+      File.join(File.dirname(__FILE__), "../fixtures/#{filepath}")
+    )
+  end #path_to
+
+  def geometry_type_for(runner)
+    result      = runner.results.first
+    table_name  = result.fetch(:tables).first
+    schema      = result.fetch(:schema)
+
+    runner.db[%Q{
+      SELECT public.GeometryType(the_geom)
+      FROM "#{schema}"."#{table_name}"
+    }].first.fetch(:geometrytype)
+  end #geometry_type_for
+end # KML regression tests
+